// src/pages/app/FormBuilder/enum.ts
export enum FieldType {
  Text = 'text',
  LongText = 'textarea',
  Email = 'email',
  Phone = 'phone',
  SingleSelect = 'singleselect',
  MultiSelect = 'multiselect',
  Radio = 'radio',
  Checkbox = 'checkbox',
  Number = 'number',
  DateTime = 'datetime',
  Date = 'date',
  Time = 'time',
  Rating = 'rating',
  File = 'file',
}

export enum DefaultFieldTypes {
  Name = 'name',
  Email = 'email',
  Phone = 'phone',
  Category = 'category',
  Organization = 'organization',
  Referral = 'referral',
  District = 'district',
  TeamId = 'team_id',
}